Terrain Demo scene created
Created foundation for terrain demo, can alter some values in runtime but should try to look at updating it so only changes mesh on slider change

+++ Assets/Scripts/Generators/Terrain Generation/MeshTerrainGenerator.cs
@@ -9,7 +9,7 @@ public class MeshTerrainGenerator : MonoBehaviour {
 	[Header("Mesh dimensions")]
 	private readonly int meshSize = 100;
 	private readonly float scale = 2f;
-	private readonly int height = 85;
+	private readonly float height = 85;
 
 	float[,] noiseMap;
